Extract `MethodCacheProxyHandler` out of `TargetHandler`.
...for a new use case, coming soon!
